Rename Counter file for consistency
